moves forms to form file, cleans up duplicate views
