+++ project_electron/orgs/form.py
@@ -1,4 +1,6 @@
 from django import forms
+from django.contrib.auth.forms import PasswordChangeForm
+
 from orgs.models import User
 
 class OrgUserUpdateForm(forms.ModelForm):
@@ -16,3 +18,22 @@ class RACSuperUserUpdateForm(forms.ModelForm):
 		model = User
 		# NO ORG -- SET TO PRIMARY
 		fields = ['is_active','email','is_superuser']
+
+class UserPasswordChangeForm(PasswordChangeForm):
+    error_css_class = 'has-error'
+    error_messages = {'password_incorrect': "You entered your current password incorrectly"}
+    old_password = forms.CharField(required=True, label='Current Password',
+                  widget=forms.PasswordInput(attrs={
+                    'class': 'form-control'}),
+                  error_messages={
+                    'required': 'Please enter your current password'})
+    new_password1 = forms.CharField(required=True, label='New Password',
+                  widget=forms.PasswordInput(attrs={
+                    'class': 'form-control'}),
+                  error_messages={
+                    'required': 'Please enter your new password'})
+    new_password2 = forms.CharField(required=True, label='New Password (repeat)',
+                  widget=forms.PasswordInput(attrs={
+                    'class': 'form-control'}),
+                  error_messages={
+                    'required': 'Please confirm your new password'})
