+++ project_electron/rac_user/form.py
@@ -1,0 +1,23 @@
+from django import forms
+from django.contrib.auth.forms import PasswordResetForm, SetPasswordForm
+
+from orgs.models import User
+
+class UserPasswordResetForm(PasswordResetForm):
+    email = forms.EmailField(required=True,
+                  widget=forms.EmailInput(attrs={
+                    'class': 'form-control has-feedback'}),
+                  error_messages={
+                    'required': 'Please enter your email'})
+
+class UserPasswordResetConfirmForm(SetPasswordForm):
+    new_password1 = forms.CharField(required=True, label='New Password',
+                  widget=forms.PasswordInput(attrs={
+                    'class': 'form-control'}),
+                  error_messages={
+                    'required': 'Please enter your new password'})
+    new_password2 = forms.CharField(required=True, label='New Password (repeat)',
+                  widget=forms.PasswordInput(attrs={
+                    'class': 'form-control'}),
+                  error_messages={
+                    'required': 'Please confirm your new password'})
